chore: moved simulation components to separate file
